# services/inventory/src/main/java/com/petek/inventory_service/entity/Product.java
package com.petek.inventory_service.entity;

import jakarta.persistence.Column;
import jakarta.persistence.Entity;
import jakarta.persistence.GeneratedValue;
import jakarta.persistence.GenerationType;
import jakarta.persistence.Id;
import jakarta.persistence.Table;
import java.math.BigDecimal;
import java.time.LocalDateTime;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@AllArgsConstructor
@NoArgsConstructor
@Builder
@Getter
@Setter
@Entity
@Table(name = "products")
public class Product {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    @Column(name = "product_id")
    private Long productId;
    
    @Column(name = "product_name")
    private String productName;
    
    private String description;
    private String category;
    
    @Column(name = "unit_of_measure")
    private String unitOfMeasure;
    
    @Column(name = "safety_stock")
    private BigDecimal safetyStock;
    
    @Column(name = "reorder_point") 
    private BigDecimal reorderPoint;
    
    @Column(name = "current_price")
    private BigDecimal currentPrice;
    
    @Column(name = "created_at")
    private LocalDateTime createdAt;
    
    @Column(name = "updated_at")
    private LocalDateTime updatedAt;
}